erstelle Dateien für Untermodule
