+++ src/gleis/gleise/daten/v3/weiche.rs
@@ -1,0 +1,8 @@
+//! Serialisierbare Darstellung von Weichen in Version 3.
+
+pub mod dreiwege;
+pub mod gerade;
+pub mod kurve;
+pub mod orientierung;
+pub mod s_kurve;
+pub mod steuerung;
